fix(clerk-js): Update the height of a form field feedback message on mount
This resolves the followin UI bug. When an error message is mounted and only requires one line for the message to be seen and then we quickly change the password to be empty. We need to update the height of the container in order for a longer message to be displayed now.